Move Threshold defs to ParseResults
